GIF reading bug fix: When reading a giant GIF, don't overdetect errors.
When reading a giant GIF, we can exhaust the possible space of codes.  At
that point we set next_code to clear_code, and store, essentially, garbage
in the clear_code position (which is never read).  But in this situation,
then, we often have code > next_code -- and this is not an error!

Bug reported by Jernej Simončič.