Add --fallback-delay option for GIFs with delay == 0
The --min-delay option raises the delay floor, but also raises the delay
of gifs with purposefully low values.

The --fallback-delay option allows setting a default value for GIFs that
do not specify a delay value without raising the minimum delay for _all_
GIFs displayed.

The --min-delay option takes precedence over the fallback value.