Two major improvements to color reduction.
1. Use gamma correction. Correct the gamma for each component using
the sRGB transfer function. Use gamma-corrected colors whenever
colors are combined or distances are measured, i.e., in adaptive
quantization and dithering. This produces reduced color palettes
much closer to input images, with more vivid appearance and stronger
light-dark contrasts. Add --gamma option for control too.

2. In the diversity algorithm, when dithering is on, consider dithered
color combinations as well as solid colors. I.e., if a color is close
to two already-selected colors in dithered combination, then avoid that
color in favor of others that cannot be so approximated.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>