Fix integer overflow by casting to unsigned as appropriate.
Recall the problem is that some variables (e.g., `uint16_t gfi->width`)
get promoted to int, and that very large GIFs might have signed 32-bit
sizes. The previous fix (just set gfi->width/height to unsigned) led
to lots of signedness warnings. We want to avoid the warnings. Really
what should be done is change all position variables to unsigned.
But I tried that, and it led to test failures. Scared of debugging the
failures, back off and fix the problem in a more targeted way.

Really a set of GIFs should be processed with clang's integer undefined
behavior sanitizer.

Conflicts:
	src/gifread.c
	src/quantize.c