Correct out-of-bounds memory access (#168).
Given a specific set of sizes, we did a floating-point multiplication
and rounding using a double in one code path and a float version of
the same factor in another. But the precision loss in the float
version led to a differently-rounded result.

So store all scaling factors as floats, not doubles, and for good
practice, don't replicate a computation, instead call the function
that contains the canonical computation.