Fix #169
The problem was that Gifsicle performed the crop on data, but left all
frames positioned into the screen by the original offset.

In other words, if the top-left-most nontransparent pixel was at
location (3, 3), then gifsicle produced an image with all frames
positioned at (3, 3) at least.

Better to position them at (0, 0) as people expect.